trying to fix login and logout
